refactor: S12.22 expose UdpSender error enum + source publicly
Move the enum SolidSyslogUdpSenderErrors and extern declaration of
UdpSenderErrorSource from SolidSyslogUdpSenderPrivate.h (library-
internal) to a new public header SolidSyslogUdpSenderErrors.h. The
integrator's error handler needs both — pointer identity on the source
to identify the emitting module, and the enum constants to switch per-
error — and Private.h was unreachable to integrator code, making the
new dispatch API unusable for the very monitoring use case it was
designed to support.

The four-TU layout becomes:
  Public:   SolidSyslogUdpSender.h       (lifecycle)
            SolidSyslogUdpSenderErrors.h (integrator error contract)
  Private:  SolidSyslogUdpSenderPrivate.h (internal struct + sigs)
            SolidSyslogUdpSender.c / Static.c (call sites)
            SolidSyslogUdpSenderMessages.c   (descriptor + AsString)

This is the pattern to replicate across all sources in phase 3.

CLAUDE.md public-headers table gains a row for the new header.
