+++ Core/Interface/SolidSyslogUdpSenderErrors.h
@@ -1,0 +1,26 @@
+#ifndef SOLIDSYSLOGUDPSENDERERRORS_H
+#define SOLIDSYSLOGUDPSENDERERRORS_H
+
+#include "ExternC.h"
+#include "SolidSyslogError.h"
+
+EXTERN_C_BEGIN
+
+    enum SolidSyslogUdpSenderErrors
+    {
+        UDPSENDER_ERROR_NULL_CONFIG,
+        UDPSENDER_ERROR_NULL_RESOLVER,
+        UDPSENDER_ERROR_NULL_DATAGRAM,
+        UDPSENDER_ERROR_NULL_ADDRESS,
+        UDPSENDER_ERROR_NULL_ENDPOINT,
+        UDPSENDER_ERROR_SEND_NULL_BUFFER,
+        UDPSENDER_ERROR_POOL_EXHAUSTED,
+        UDPSENDER_ERROR_UNKNOWN_DESTROY,
+        UDPSENDER_ERROR_MAX
+    };
+
+    extern const struct SolidSyslogErrorSource UdpSenderErrorSource;
+
+EXTERN_C_END
+
+#endif /* SOLIDSYSLOGUDPSENDERERRORS_H */
