+++ Tests/SolidSyslogUdpSenderTest.cpp
@@ -14,8 +14,9 @@
 #include "SolidSyslogPosixDatagram.h"
 #include "SolidSyslogTunables.h"
 #include "SolidSyslogUdpSender.h"
-#include "SolidSyslogUdpSenderPrivate.h"
+#include "SolidSyslogUdpSenderErrors.h"
 #include "SolidSyslogSender.h"
+#include "SolidSyslogSenderDefinition.h"
 #include "SocketFake.h"
 #include "SolidSyslogDatagram.h"
 #include "TestUtils.h"
